Fixed tests to accomodate new context requirements and unit system serialization
