+++ examples/dev/dev_validate_params.py
@@ -1,7 +1,7 @@
 import os
 import re
 
-from flow360.component.flow360_params.flow360_legacy import Flow360ParamsLegacy
+from flow360.component.flow360_params.flow360_params import Flow360ParamsLegacy
 
 rootdir = "../../tests/data/cases/"
 regex = re.compile("(case_.*\.json$)")
