+++ flow360/component/flow360_params/flow360_output.py
@@ -9,6 +9,7 @@ from typing import List, Literal, Optional, Union, get_args
 import pydantic as pd
 from pydantic import conlist
 
+from .flow360_legacy import LegacyModel, _get_output_fields
 from ..types import Coordinate, PositiveInt
 from .flow360_fields import (
     CommonFieldNames,
@@ -462,3 +463,155 @@ class AeroacousticOutput(Flow360BaseModel):
     patch_type: Optional[str] = pd.Field("solid", const=True, alias="patchType")
     observers: List[Coordinate] = pd.Field()
     write_per_surface_output: Optional[bool] = pd.Field(False, alias="writePerSurfaceOutput")
+
+
+class LegacyOutputFormat(pd.BaseModel, metaclass=ABCMeta):
+    """:class: Base class for common output parameters"""
+
+    Cp: Optional[bool] = pd.Field()
+    grad_w: Optional[bool] = pd.Field(alias="gradW")
+    k_omega: Optional[bool] = pd.Field(alias="kOmega")
+    Mach: Optional[bool] = pd.Field(alias="Mach")
+    mut: Optional[bool] = pd.Field()
+    mut_ratio: Optional[bool] = pd.Field(alias="mutRatio")
+    nu_hat: Optional[bool] = pd.Field(alias="nuHat")
+    primitive_vars: Optional[bool] = pd.Field(alias="primitiveVars")
+    q_criterion: Optional[bool] = pd.Field(alias="qcriterion")
+    residual_navier_stokes: Optional[bool] = pd.Field(alias="residualNavierStokes")
+    residual_transition: Optional[bool] = pd.Field(alias="residualTransition")
+    residual_turbulence: Optional[bool] = pd.Field(alias="residualTurbulence")
+    s: Optional[bool] = pd.Field()
+    solution_navier_stokes: Optional[bool] = pd.Field(alias="solutionNavierStokes")
+    solution_turbulence: Optional[bool] = pd.Field(alias="solutionTurbulence")
+    solution_transition: Optional[bool] = pd.Field(alias="solutionTransition")
+    T: Optional[bool] = pd.Field(alias="T")
+    vorticity: Optional[bool] = pd.Field()
+    wall_distance: Optional[bool] = pd.Field(alias="wallDistance")
+    low_numerical_dissipation_sensor: Optional[bool] = pd.Field(
+        alias="lowNumericalDissipationSensor"
+    )
+    residual_heat_solver: Optional[bool] = pd.Field(alias="residualHeatSolver")
+
+
+# pylint: disable=too-many-ancestors
+class SurfaceOutputLegacy(SurfaceOutput, LegacyOutputFormat, LegacyModel):
+    """:class:`SurfaceOutputLegacy` class"""
+
+    wall_function_metric: Optional[bool] = pd.Field(alias="wallFunctionMetric")
+    node_moments_per_unit_area: Optional[bool] = pd.Field(alias="nodeMomentsPerUnitArea")
+    residual_sa: Optional[bool] = pd.Field(alias="residualSA")
+    coarsen_iterations: Optional[int] = pd.Field(alias="coarsenIterations")
+
+    Cf: Optional[bool] = pd.Field(alias="Cf")
+    Cf_vec: Optional[bool] = pd.Field(alias="CfVec")
+    Cf_normal: Optional[bool] = pd.Field(alias="CfNormal")
+    Cf_tangent: Optional[bool] = pd.Field(alias="CfTangent")
+    y_plus: Optional[bool] = pd.Field(alias="yPlus")
+    wall_distance: Optional[bool] = pd.Field(alias="wallDistance")
+    heat_flux: Optional[bool] = pd.Field(alias="heatFlux")
+    node_forces_per_unit_area: Optional[bool] = pd.Field(alias="nodeForcesPerUnitArea")
+    node_normals: Optional[bool] = pd.Field(alias="nodeNormals")
+    velocity_relative: Optional[bool] = pd.Field(alias="VelocityRelative")
+
+    def update_model(self) -> Flow360BaseModel:
+        fields = _get_output_fields(self, [])
+
+        if self.output_fields is not None:
+            fields += self.output_fields
+
+        model = {
+            "animationSettings": {
+                "frequency": self.animation_frequency,
+                "frequencyOffset": self.animation_frequency_offset,
+                "frequencyTimeAverage": self.animation_frequency_time_average,
+                "frequencyTimeAverageOffset": self.animation_frequency_time_average_offset,
+            },
+            "computeTimeAverages": self.compute_time_averages,
+            "outputFormat": self.output_format,
+            "outputFields": fields,
+            "startAverageIntegrationStep": self.start_average_integration_step,
+        }
+
+        return SurfaceOutput.parse_obj(model)
+
+
+class SliceOutputLegacy(SliceOutput, LegacyOutputFormat, LegacyModel):
+    """:class:`SliceOutputLegacy` class"""
+
+    coarsen_iterations: Optional[int] = pd.Field(alias="coarsenIterations")
+    bet_metrics: Optional[bool] = pd.Field(alias="betMetrics")
+    bet_metrics_per_disk: Optional[bool] = pd.Field(alias="betMetricsPerDisk")
+
+    def update_model(self) -> Flow360BaseModel:
+        fields = _get_output_fields(self, [])
+
+        if self.output_fields is not None:
+            fields += self.output_fields
+
+        model = {
+            "animationSettings": {
+                "frequency": self.animation_frequency,
+                "frequencyOffset": self.animation_frequency_offset,
+            },
+            "outputFormat": self.output_format,
+            "outputFields": fields,
+            "slices": self.slices,
+        }
+
+        return SliceOutput.parse_obj(model)
+
+
+# pylint: disable=too-many-ancestors
+class VolumeOutputLegacy(VolumeOutput, LegacyOutputFormat, LegacyModel):
+    """:class:`VolumeOutputLegacy` class"""
+
+    write_single_file: Optional[bool] = pd.Field(alias="writeSingleFile")
+    write_distributed_file: Optional[bool] = pd.Field(alias="writeDistributedFile")
+    residual_components_sa: Optional[bool] = pd.Field(alias="residualComponentsSA")
+    wall_distance_dir: Optional[bool] = pd.Field(alias="wallDistanceDir")
+    velocity_relative: Optional[bool] = pd.Field(alias="VelocityRelative")
+    debug_transition: Optional[bool] = pd.Field(alias="debugTransition")
+    debug_turbulence: Optional[bool] = pd.Field(alias="debugTurbulence")
+    debug_navier_stokes: Optional[bool] = pd.Field(alias="debugNavierStokes")
+    bet_metrics: Optional[bool] = pd.Field(alias="betMetrics")
+    bet_metrics_per_disk: Optional[bool] = pd.Field(alias="betMetricsPerDisk")
+
+    def update_model(self) -> Flow360BaseModel:
+        fields = _get_output_fields(self, ["write_single_file", "write_distributed_file"])
+
+        if self.output_fields is not None:
+            fields += self.output_fields
+
+        model = {
+            "animationSettings": {
+                "frequency": self.animation_frequency,
+                "frequencyOffset": self.animation_frequency_offset,
+                "frequencyTimeAverage": self.animation_frequency_time_average,
+                "frequencyTimeAverageOffset": self.animation_frequency_time_average_offset,
+            },
+            "computeTimeAverages": self.compute_time_averages,
+            "outputFormat": self.output_format,
+            "outputFields": fields,
+            "startAverageIntegrationStep": self.start_average_integration_step,
+        }
+
+        return VolumeOutput.parse_obj(model)
+
+
+""" Legacy models for Flow360 updater, do not expose """
+
+
+class IsoSurfaceOutputLegacy(IsoSurfaceOutput, LegacyModel):
+    """:class:`IsoSurfaceOutputLegacy` class"""
+
+    def update_model(self):
+        model = {
+            "animationSettings": {
+                "frequency": self.animation_frequency,
+                "frequencyOffset": self.animation_frequency_offset,
+            },
+            "outputFormat": self.output_format,
+            "isoSurfaces": self.iso_surfaces,
+        }
+
+        return IsoSurfaceOutput.parse_obj(model)
