+++ flow360/component/flow360_params/solvers.py
@@ -9,6 +9,7 @@ from typing import List, Optional, Union
 import pydantic as pd
 from typing_extensions import Literal
 
+from .flow360_legacy import LegacyModel, _try_update, _try_set
 from ..types import NonNegativeFloat, NonNegativeInt, PositiveFloat, PositiveInt
 from .params_base import DeprecatedAlias, Flow360BaseModel
 
@@ -376,3 +377,171 @@ class TransitionModelSolver(GenericFlowSolverSettings):
         alias="turbulenceIntensityPercent"
     )
     N_crit: Optional[PositiveFloat] = pd.Field(alias="Ncrit")
+
+
+""" Legacy models for Flow360 updater, do not expose """
+
+class LinearSolverLegacy(LinearSolver, LegacyModel):
+    """:class:`LinearSolverLegacy` class"""
+
+    max_level_limit: Optional[NonNegativeInt] = pd.Field(alias="maxLevelLimit")
+
+    def update_model(self) -> Flow360BaseModel:
+        model = {
+            "absoluteTolerance": self.absolute_tolerance,
+            "relativeTolerance": self.relative_tolerance,
+            "maxIterations": self.max_iterations,
+        }
+
+        return LinearSolver.parse_obj(model)
+
+
+class PressureCorrectionSolverLegacy(PressureCorrectionSolver, LegacyModel):
+    """:class:`PressureCorrectionSolverLegacy` class"""
+
+    linear_solver: Optional[LinearSolverLegacy] = pd.Field(
+        alias="linearSolver", default=LinearSolverLegacy()
+    )
+
+    def update_model(self) -> Flow360BaseModel:
+        model = {"randomizer": self.randomizer, "linear_solver": _try_update(self.linear_solver)}
+
+        return PressureCorrectionSolver.parse_obj(model)
+
+
+class NavierStokesSolverLegacy(NavierStokesSolver, LegacyModel):
+    """:class:`NavierStokesModelSolverLegacy` class"""
+
+    viscous_wave_speed_scale: Optional[float] = pd.Field(alias="viscousWaveSpeedScale")
+    extra_dissipation: Optional[float] = pd.Field(alias="extraDissipation")
+    pressure_correction_solver: Optional[PressureCorrectionSolver] = pd.Field(
+        alias="pressureCorrectionSolver"
+    )
+    randomizer: Optional[LinearIterationsRandomizer] = pd.Field()
+    linear_solver: Optional[LinearSolverLegacy] = pd.Field(
+        alias="linearSolver", default=LinearSolverLegacy()
+    )
+
+    def update_model(self) -> Flow360BaseModel:
+        model = {
+            "absoluteTolerance": self.absolute_tolerance,
+            "relativeTolerance": self.relative_tolerance,
+            "CFLMultiplier": self.CFL_multiplier,
+            "linearSolver": _try_update(self.linear_solver),
+            "updateJacobianFrequency": self.update_jacobian_frequency,
+            "equationEvalFrequency": self.equation_eval_frequency,
+            "maxForceJacUpdatePhysicalSteps": self.max_force_jac_update_physical_steps,
+            "orderOfAccuracy": self.order_of_accuracy,
+            "kappaMUSCL": self.kappa_MUSCL,
+            "limitVelocity": self.limit_velocity,
+            "limitPressureDensity": self.limit_pressure_density,
+            "numericalDissipationFactor": self.numerical_dissipation_factor,
+        }
+
+        if self.linear_iterations is not None and model["linearSolver"] is not None:
+            model["linearSolver"].max_iterations = self.linear_iterations
+
+        return NavierStokesSolver.parse_obj(model)
+
+
+class TurbulenceModelSolverLegacy(TurbulenceModelSolver, LegacyModel):
+    """:class:`TurbulenceModelSolverLegacy` class"""
+
+    randomizer: Optional[LinearIterationsRandomizer] = pd.Field()
+    kappa_MUSCL: Optional[pd.confloat(ge=-1, le=1)] = pd.Field(alias="kappaMUSCL")
+    linear_iterations: Optional[PositiveInt] = pd.Field(alias="linearIterations")
+    linear_solver: Optional[LinearSolverLegacy] = pd.Field(
+        alias="linearSolver", default=LinearSolverLegacy()
+    )
+    rotation_correction: Optional[bool] = pd.Field(alias="rotationCorrection")
+
+    def update_model(self) -> Flow360BaseModel:
+        class _TurbulenceTempModel(pd.BaseModel):
+            """Helper class used to create
+            the correct solver from dict data"""
+
+            solver: TurbulenceModelSolverTypes = pd.Field(discriminator="model_type")
+
+        model = {
+            "solver": {
+                "absoluteTolerance": self.absolute_tolerance,
+                "relativeTolerance": self.relative_tolerance,
+                "modelType": self.model_type,
+                "CFLMultiplier": self.CFL_multiplier,
+                "linearSolver": _try_update(self.linear_solver),
+                "updateJacobianFrequency": self.update_jacobian_frequency,
+                "equationEvalFrequency": self.equation_eval_frequency,
+                "maxForceJacUpdatePhysicalSteps": self.max_force_jac_update_physical_steps,
+                "orderOfAccuracy": self.order_of_accuracy,
+                "DDES": self.DDES,
+                "gridSizeForLES": self.grid_size_for_LES,
+                "quadraticConstitutiveRelation": self.quadratic_constitutive_relation,
+                "reconstructionGradientLimiter": self.reconstruction_gradient_limiter,
+                "modelConstants": self.model_constants,
+            }
+        }
+
+        if isinstance(self, TurbulenceModelSolverSA):
+            _try_set(model, "rotationCorrection", self.rotation_correction)
+
+        if self.linear_iterations is not None and model["solver"]["linearSolver"] is not None:
+            model["solver"]["linearSolver"].max_iterations = self.linear_iterations
+
+        return _TurbulenceTempModel.parse_obj(model).solver
+
+
+class HeatEquationSolverLegacy(HeatEquationSolver, LegacyModel):
+    """:class:`HeatEquationSolverLegacy` class"""
+
+    order_of_accuracy: Optional[Literal[1, 2]] = pd.Field(alias="orderOfAccuracy")
+    model_type: Literal["HeatEquation"] = pd.Field("HeatEquation", alias="modelType")
+    CFL_multiplier: Optional[PositiveFloat] = pd.Field(alias="CFLMultiplier")
+    update_jacobian_frequency: Optional[PositiveInt] = pd.Field(alias="updateJacobianFrequency")
+    max_force_jac_update_physical_steps: Optional[NonNegativeInt] = pd.Field(
+        alias="maxForceJacUpdatePhysicalSteps"
+    )
+    linear_solver: Optional[LinearSolverLegacy] = pd.Field(
+        alias="linearSolver", default=LinearSolverLegacy()
+    )
+
+    def update_model(self) -> Flow360BaseModel:
+        model = {
+            "absoluteTolerance": self.absolute_tolerance,
+            "linearSolver": _try_update(self.linear_solver),
+            "equationEvalFrequency": self.equation_eval_frequency,
+        }
+
+        return HeatEquationSolver.parse_obj(model)
+
+
+class TransitionModelSolverLegacy(TransitionModelSolver, LegacyModel):
+    """:class:`TransitionModelSolverLegacy` class"""
+
+    reconstruction_gradient_limiter: Optional[pd.confloat(ge=0, le=2)] = pd.Field(
+        alias="reconstructionGradientLimiter"
+    )
+    CFL_multiplier: Optional[PositiveFloat] = pd.Field(alias="CFLMultiplier")
+    linear_iterations: Optional[PositiveInt] = pd.Field(alias="linearIterations")
+    randomizer: Optional[LinearIterationsRandomizer] = pd.Field()
+    linear_solver: Optional[LinearSolverLegacy] = pd.Field(
+        alias="linearSolver", default=LinearSolverLegacy()
+    )
+
+    def update_model(self) -> Flow360BaseModel:
+        model = {
+            "absoluteTolerance": self.absolute_tolerance,
+            "relativeTolerance": self.relative_tolerance,
+            "modelType": self.model_type,
+            "linearSolver": _try_update(self.linear_solver),
+            "updateJacobianFrequency": self.update_jacobian_frequency,
+            "equationEvalFrequency": self.equation_eval_frequency,
+            "maxForceJacUpdatePhysicalSteps": self.max_force_jac_update_physical_steps,
+            "orderOfAccuracy": self.order_of_accuracy,
+            "turbulenceIntensityPercent": self.turbulence_intensity_percent,
+            "Ncrit": self.N_crit,
+        }
+
+        if self.linear_iterations is not None and model["linearSolver"] is not None:
+            model["linearSolver"].max_iterations = self.linear_iterations
+
+        return HeatEquationSolver.parse_obj(model)
