+++ tests/om6wing/Flow360Mesh.json
@@ -1,0 +1,6 @@
+{
+    "boundaries" :
+    {
+        "noSlipWalls" : [1]
+    }
+}
