+++ tests/params/test_solvers.py
@@ -2,6 +2,7 @@ import unittest
 
 import pydantic as pd
 import pytest
+import flow360 as fl
 
 from flow360.component.flow360_params.flow360_params import (
     Flow360Params,
@@ -13,7 +14,7 @@ from flow360.component.flow360_params.flow360_params import (
     TurbulenceModelSolverSA,
     TurbulenceModelSolverSST,
 )
-from tests.utils import compare_to_ref, to_file_from_file_test
+from tests.utils import compare_to_ref, to_file_from_file_test, to_file_from_file_params_test
 
 assertions = unittest.TestCase("__init__")
 
@@ -53,11 +54,12 @@ def test_navier_stokes():
         limit_velocity=True,
         limit_pressure_density=False,
     )
-    p = Flow360Params(
-        navier_stokes_solver=ns,
-        freestream={"Mach": 1, "Temperature": 1, "muRef": 1},
-    )
-    to_file_from_file_test(p)
+    with fl.SI_unit_system:
+        p = Flow360Params(
+            navier_stokes_solver=ns,
+            freestream={"Mach": 1, "Temperature": 1, "muRef": 1},
+        )
+    to_file_from_file_params_test(p)
 
 
 def test_turbulence_solver():
