+++ tests/ref/case_params/params.json
@@ -1,35 +1,81 @@
 {
+    "unitSystem": {
+        "mass": {
+            "value": 1.0,
+            "units": "kg"
+        },
+        "length": {
+            "value": 1.0,
+            "units": "m"
+        },
+        "time": {
+            "value": 1.0,
+            "units": "s"
+        },
+        "temperature": {
+            "value": 1.0,
+            "units": "K"
+        },
+        "velocity": {
+            "value": 1.0,
+            "units": "m/s"
+        },
+        "area": {
+            "value": 1.0,
+            "units": "m**2"
+        },
+        "force": {
+            "value": 1.0,
+            "units": "N"
+        },
+        "pressure": {
+            "value": 1.0,
+            "units": "Pa"
+        },
+        "density": {
+            "value": 1.0,
+            "units": "kg/m**3"
+        },
+        "viscosity": {
+            "value": 1.0,
+            "units": "Pa*s"
+        },
+        "angular_velocity": {
+            "value": 1.0,
+            "units": "rad/s"
+        }
+    },
     "version": "0.2.0b16",
     "geometry": {
         "refArea": {
-            "units": "m**2",
-            "value": 1.15315084119231
+            "value": 1.15315084119231,
+            "units": "m**2"
         },
         "momentLength": {
-            "units": "m",
             "value": [
                 1.47602,
                 0.801672958512342,
                 1.47602
-            ]
+            ],
+            "units": "m"
         },
         "meshUnit": {
-            "units": "m",
-            "value": 1.0
+            "value": 1.0,
+            "units": "m"
         }
     },
     "boundaries": {
-        "1": {
-            "type": "NoSlipWall",
-            "name": "wing"
+        "2": {
+            "type": "SlipWall",
+            "name": "symmetry"
         },
         "3": {
             "type": "Freestream",
             "name": "freestream"
         },
-        "2": {
-            "type": "SlipWall",
-            "name": "symmetry"
+        "1": {
+            "type": "NoSlipWall",
+            "name": "wing"
         }
     },
     "timeStepping": {
@@ -37,47 +83,29 @@
     },
     "navierStokesSolver": {
         "linearIterations": 10,
-        "linearSolver":
-        {
-            "absoluteTolerance": 1e-10,
-            "maxIterations": 50
+        "linearSolver": {
+            "maxIterations": 50,
+            "absoluteTolerance": 1e-10
         }
     },
     "freestream": {
+        "alphaAngle": 3.06,
+        "betaAngle": 0.0,
         "velocity": {
             "value": 286.0,
             "units": "m/s"
-        },
-        "alphaAngle": 3.06,
-        "betaAngle": 0.0
+        }
     },
     "surfaceOutput": {
+        "animationSettings": {},
         "outputFormat": "tecplot",
         "outputFields": [
             "yPlus",
             "primitiveVars"
-        ],
-        "surfaces": {
-            "freestream": {
-                "outputFields": [
-                    "Cp",
-                    "Mach"
-                ]
-            },
-            "symmetry": {
-                "outputFields": [
-                    "T",
-                    "heatFlux"
-                ]
-            },
-            "wing": {
-                "outputFields": [
-                    "CfVec"
-                ]
-            }
-        }
+        ]
     },
     "volumeOutput": {
+        "animationSettings": {},
         "outputFormat": "tecplot",
         "outputFields": [
             "Mach",
@@ -85,42 +113,43 @@
         ]
     },
     "sliceOutput": {
+        "animationSettings": {},
         "outputFormat": "tecplot",
         "outputFields": [
             "Cp"
         ],
         "slices": {
             "x0": {
-                "outputFields": [
-                    "Mach"
-                ],
                 "sliceNormal": [
-                    1,
-                    0,
-                    0
+                    1.0,
+                    0.0,
+                    0.0
                 ],
                 "sliceOrigin": [
-                    0,
-                    0,
-                    0
+                    0.0,
+                    0.0,
+                    0.0
+                ],
+                "outputFields": [
+                    "Mach"
                 ]
             },
             "y1": {
-                "outputFields": [
-                    "T"
-                ],
                 "sliceNormal": [
-                    0,
-                    1,
-                    0
+                    0.0,
+                    1.0,
+                    0.0
                 ],
                 "sliceOrigin": [
-                    2,
-                    1,
-                    0
+                    2.0,
+                    1.0,
+                    0.0
+                ],
+                "outputFields": [
+                    "T"
                 ]
             }
         }
     },
-    "hash": "3bc357f0032849255fa4ef5f52932d74486ee3417383ba32f80ab8a1ac345a93"
+    "hash": "a4ffb6ec1c6d5001b5281d45129a74fb1f9a9973524831ff1c376a75d72e9a70"
 }
