+++ tests/ref/case_params/params_units_solver.json
@@ -1,102 +1,102 @@
 {
-    "geometry": {
-        "refArea": 1.0,
-        "momentCenter": [
-            1,
-            2,
-            3
-        ],
-        "momentLength": [
-            37.490908,
-            20.362493146213485,
-            37.490908
-        ]
+  "geometry": {
+    "refArea": 1.0,
+    "momentCenter": [
+      1,
+      2,
+      3
+    ],
+    "momentLength": [
+      37.490908,
+      20.362493146213485,
+      37.490908
+    ]
+  },
+  "boundaries": {
+    "2": {
+      "type": "SlipWall",
+      "name": "symmetry"
     },
-    "boundaries": {
-        "2": {
-            "type": "SlipWall",
-            "name": "symmetry"
-        },
-        "1": {
-            "type": "NoSlipWall",
-            "name": "wing",
-            "Velocity": [
-                0.0008162876014170267,
-                0.0016325752028340534,
-                0.0024488628042510802
-            ]
-        },
-        "3": {
-            "type": "Freestream",
-            "name": "freestream"
-        }
+    "1": {
+      "type": "NoSlipWall",
+      "name": "wing",
+      "Velocity": [
+        0.0008162876014170267,
+        0.0016325752028340534,
+        0.0024488628042510802
+      ]
     },
-    "timeStepping": {
-        "maxPseudoSteps": 500,
-        "timeStepSize": 408352.8069698554,
-        "CFL": {
-            "type": "adaptive",
-            "min": 0.1,
-            "max": 10000.0,
-            "maxRelativeChange": 1.0,
-            "convergenceLimitingFactor": 0.25
-        }
-    },
-    "freestream": {
-        "alphaAngle": 3.06,
-        "betaAngle": 0.0,
-        "Mach": 0.8404497144189705,
-        "muRef": 4.292519319815164e-05,
-        "Temperature": 288.15
-    },
-    "volumeZones": {
-        "zone2": {
-            "modelType": "FluidDynamics",
-            "referenceFrame": {
-                "omegaRadians": 3.0773317581937964e-06,
-                "centerOfRotation": [
-                    0.0,
-                    0.0,
-                    0.0
-                ],
-                "axisOfRotation": [
-                    1,
-                    0,
-                    0
-                ]
-            }
-        },
-        "zone1": {
-            "modelType": "FluidDynamics",
-            "referenceFrame": {
-                "omegaRadians": 3.0773317581937964e-06,
-                "centerOfRotation": [
-                    0.0,
-                    0.0,
-                    0.0
-                ],
-                "axisOfRotation": [
-                    1,
-                    0,
-                    0
-                ]
-            }
-        },
-        "zone3": {
-            "modelType": "FluidDynamics",
-            "referenceFrame": {
-                "omegaRadians": 3.077331758193797e-06,
-                "centerOfRotation": [
-                    0.0,
-                    0.0,
-                    0.0
-                ],
-                "axisOfRotation": [
-                    1,
-                    0,
-                    0
-                ]
-            }
-        }
+    "3": {
+      "type": "Freestream",
+      "name": "freestream"
     }
+  },
+  "timeStepping": {
+    "maxPseudoSteps": 500,
+    "timeStepSize": 408352.8069698554,
+    "CFL": {
+      "type": "adaptive",
+      "min": 0.1,
+      "max": 10000.0,
+      "maxRelativeChange": 1.0,
+      "convergenceLimitingFactor": 0.25
+    }
+  },
+  "freestream": {
+    "alphaAngle": 3.06,
+    "betaAngle": 0.0,
+    "Mach": 0.8404497144189705,
+    "muRef": 4.292519319815164e-05,
+    "Temperature": 288.15
+  },
+  "volumeZones": {
+    "zone3": {
+      "modelType": "FluidDynamics",
+      "referenceFrame": {
+        "omegaRadians": 3.077331758193797e-06,
+        "centerOfRotation": [
+          0.0,
+          0.0,
+          0.0
+        ],
+        "axisOfRotation": [
+          1,
+          0,
+          0
+        ]
+      }
+    },
+    "zone2": {
+      "modelType": "FluidDynamics",
+      "referenceFrame": {
+        "omegaRadians": 3.0773317581937964e-06,
+        "centerOfRotation": [
+          0.0,
+          0.0,
+          0.0
+        ],
+        "axisOfRotation": [
+          1,
+          0,
+          0
+        ]
+      }
+    },
+    "zone1": {
+      "modelType": "FluidDynamics",
+      "referenceFrame": {
+        "omegaRadians": 3.0773317581937964e-06,
+        "centerOfRotation": [
+          0.0,
+          0.0,
+          0.0
+        ],
+        "axisOfRotation": [
+          1,
+          0,
+          0
+        ]
+      }
+    }
+  }
 }
